requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
tiktoken
pinecone-client[grpc]==2.2.4
pydantic
tenacity
termcolor
